OrganisationConverter refactored
