+++ web/src/main/resources/application.properties
@@ -5,3 +5,6 @@ management.security.enabled=false
 logging.level.org.springframework.web=DEBUG
 logging.level.org.hibernate=ERROR
 logging.file.path=web/webApplication.log
+
+server.error.include-message=always
+server.error.include-binding-errors=always
